+++ CHANGELOG.md
@@ -24,6 +24,8 @@ compatible with the updates.
 ### Changed
 
 - Changed `no-test` flag to `full` for demo scripts.
+- Renamed `neg_weight` to `background_weight`.
+- Renamed `log_dir` to `exp_name` and `log_root` to `log_dir` respectively.
 - Uniformed local-net, global-net, u-net under a single u-net structure.
 - Simplified custom layer definitions.
 - Removed multiple unnecessary custom layers and use tf.keras.layers whenever possible.